Added the Module abstract class
